+++ src/haddock/libs/libontology.py
@@ -71,6 +71,7 @@ class PDBFile(Persistent):
         topology: Optional[Any] = None,
         aa_topology: Optional[Any] = None,
         cgtoaa_tbl:Optional[Any] = None,
+        shape:Optional[Any] = False,
         path: Union[Path, str] = ".",
         score: float = NaN,
         md5: Optional[str] = None,
@@ -82,6 +83,7 @@ class PDBFile(Persistent):
         self.topology = topology
         self.aa_topology = aa_topology 
         self.cgtoaa_tbl = cgtoaa_tbl
+        self.shape = shape
         self.score = score
         self.ori_name: Optional[str] = None
         self.clt_id: Union[str, int, None] = None